fix: resolve pr review comments
